+++ sakura_core/cmd/ViewCommander_CustMenu.cpp
@@ -43,7 +43,7 @@ void ViewCommander::Command_Menu_RButton(void)
 			pszWork[nLength] = _T('\0');
 
 			// Œ©‚½–Ú‚Æ“¯‚¶‚æ‚¤‚ÉA\n ‚ğ CR+LF‚Ö•ÏŠ·‚·‚é
-			for (int i=0; i<nLength; ++i) {
+			for (size_t i=0; i<nLength; ++i) {
 				if (pszWork[i] == _T('\\') && pszWork[i + 1] == _T('n')) {
 					pszWork[i] = WCODE::CR;
 					pszWork[i+1] = WCODE::LF;
